Fix compras y  actualizacion de modelo

+++ DATA/EntityDataModel/BDDistBHF/VR_GO_DespachoXProducto.vb
@@ -22,7 +22,7 @@ Partial Public Class VR_GO_DespachoXProducto
     Public Property oaest As Nullable(Of Integer)
     Public Property oaap As Nullable(Of Integer)
     Public Property oafdoc As Nullable(Of Date)
-    Public Property caja As Nullable(Of Decimal)
+    Public Property caja As Nullable(Of Integer)
     Public Property Unidad As Nullable(Of Decimal)
     Public Property Total As Nullable(Of Decimal)
